requirements.txt
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==2.1
django-allauth==0.39.1
django-extensions==2.1.9
django-rest-auth==0.9.5
djangorestframework==3.9.4
factory-boy==2.12.0
Faker==1.0.7
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
oauthlib==3.0.2
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
text-unidecode==1.2
uritemplate==3.0.0
urllib3==1.25.3
